requirements.txt
python-telegram-bot>=20.0
aiohttp>=3.8.0
flask>=2.3.0
qrcode[pil]>=7.4
python-dotenv>=1.0.0
Pillow>=10.0.0
